reorganized dirs and added alias @->root/src
