+++ package.json
@@ -7,10 +7,11 @@
     "smui-theme-dark": "smui-theme compile public/build/smui-dark.css -i src/theme/dark",
     "build": "rollup -c",
     "dev": "rollup -c -w",
-    "start": "sirv public --no-clear",
+    "start": "sirv public --no-clear --host",
     "test": "env TS_NODE_COMPILER_OPTIONS='{\"module\":\"commonjs\",\"types\":[\"mocha\"]}' ./node_modules/.bin/mocha -r ts-node/register ./tests/*.spec.ts"
   },
   "devDependencies": {
+    "@rollup/plugin-alias": "^3.1.2",
     "@rollup/plugin-commonjs": "^17.0.0",
     "@rollup/plugin-node-resolve": "^11.0.0",
     "@rollup/plugin-typescript": "^8.2.1",
